+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Added (new features/APIs/variables/...)
 
+- [PR 124](https://github.com/athelas-astro/athelas/pull/124) Use minimum internal energy in BEL
 - [PR 123](https://github.com/athelas-astro/athelas/pull/123) Move basis, eos, opac objects into MeshState
 - [PR 122](https://github.com/athelas-astro/athelas/pull/122) Threaded StageData into BEL
 - [PR 121](https://github.com/athelas-astro/athelas/pull/121) New Saha solver; optimizations
@@ -15,6 +16,8 @@
 ### Fixed (not changing behavior/API/variables/...)
 
 ### Infrastructure (changes irrelevant to downstream codes)
+
+- [PR 124](https://github.com/athelas-astro/athelas/pull/124) Introduce `EOSLambda` lambda wrapper
 
 ### Removed (removing behavior/API/variables/...)